chore: wip on tests
