+++ package.json
@@ -27,19 +27,21 @@
     "changeset:release": "changeset publish"
   },
   "devDependencies": {
+    "@types/mock-fs": "^4.13.4",
     "@types/node": "^20.9.0",
     "@typescript-eslint/eslint-plugin": "^6.10.0",
     "@typescript-eslint/parser": "^6.10.0",
     "eslint": "^8.57.0",
     "husky": "^8.0.3",
     "lint-staged": "^15.1.0",
+    "mock-fs": "^5.2.0",
     "pkg": "^5.8.1",
     "prettier": "3.2.5",
     "rimraf": "^5.0.5",
     "ts-node": "^10.9.2",
     "tsup": "^8.0.2",
     "typescript": "^5.2.2",
-    "vitest": "^1.3.1"
+    "vitest": "^1.5.0"
   },
   "dependencies": {
     "@changesets/changelog-github": "^0.5.0",
