+++ packages/cli/src/icons/icon-name.ts
@@ -1,0 +1,3 @@
+export function iconName(file: string) {
+  return file.replace(/\.svg$/, '').replace(/\\/g, '/');
+}
