+++ packages/cli/src/icons/index.ts
@@ -1,0 +1,1 @@
+export { generateSprite } from './generate-sprite';
